Added capability to recognize which of the identical actions was triggered for Overlaybuttons
Next big task is to look into Forests and implement support for
dangling edges: constituent edges that are not connected to any node in
other end. These are structurally impossible, but needed for drawing.
They should be marked with accent color as long as they exist in the
structure, as structure is broken when there are dangling edges.
Probably visualization algorithms etc. need also verify that edges
really are connected somewhere before trying to operate with danglers.